fix(battle): int32 overflow in RE piercing-DEF damage calc; fix(char): 10min delete cooldown
battle_calc_defense_reduction() multiplied def1 * skill_ratio in plain
32-bit int before folding into the int64 damage accumulator. High
skill-ratio stacks (custom trait stats) overflowed INT32_MAX and
wrapped negative, which the client silently drops -- reported as
"damage above 1B doesn't show". Same shape fixed in
SR_CRESCENTELBOW_AUTOSPELL's ratio local.

char_del_delay 1800 -> 600 (30min -> 10min).

+++ src/map/battle.cpp
@@ -4860,8 +4860,8 @@ static void battle_calc_defense_reduction( Damage* wd, block_list* src, block_li
 		if (def1 == -400) /* -400 creates a division by 0 and subsequently crashes */
 			def1 = -399;
 		ATK_ADD2(wd->damage, wd->damage2,
-			is_attack_piercing(wd, src, target, skill_id, skill_lv, EQI_HAND_R) ? (def1 * battle_calc_attack_skill_ratio(wd, src, target, skill_id, skill_lv)) / 200 : 0,
-			is_attack_piercing(wd, src, target, skill_id, skill_lv, EQI_HAND_L) ? (def1 * battle_calc_attack_skill_ratio(wd, src, target, skill_id, skill_lv)) / 200 : 0
+			is_attack_piercing(wd, src, target, skill_id, skill_lv, EQI_HAND_R) ? ((int64)def1 * battle_calc_attack_skill_ratio(wd, src, target, skill_id, skill_lv)) / 200 : 0,
+			is_attack_piercing(wd, src, target, skill_id, skill_lv, EQI_HAND_L) ? ((int64)def1 * battle_calc_attack_skill_ratio(wd, src, target, skill_id, skill_lv)) / 200 : 0
 		);
 		if (!attack_ignores_def(wd, src, target, skill_id, skill_lv, EQI_HAND_R) && !is_attack_piercing(wd, src, target, skill_id, skill_lv, EQI_HAND_R))
 			wd->damage = wd->damage * (4000 + def1) / (4000 + 10 * def1) - vit_def;
@@ -5207,7 +5207,7 @@ static void battle_calc_weapon_final_atk_modifiers(struct Damage* wd, block_list
 	if( tsc && tsc->getSCE(SC_CRESCENTELBOW) && wd->flag&BF_SHORT && rnd()%100 < tsc->getSCE(SC_CRESCENTELBOW)->val2 ) {
 		//ATK [{(Target HP / 100) x Skill Level} x Caster Base Level / 125] % + [Received damage x {1 + (Skill Level x 0.2)}]
 		int64 rdamage = 0;
-		int32 ratio = (int64)(status_get_hp(src) / 100) * tsc->getSCE(SC_CRESCENTELBOW)->val1 * status_get_lv(target) / 125;
+		int64 ratio = (int64)(status_get_hp(src) / 100) * tsc->getSCE(SC_CRESCENTELBOW)->val1 * status_get_lv(target) / 125;
 		if (ratio > 5000) ratio = 5000; // Maximum of 5000% ATK
 		rdamage = battle_calc_base_damage(target,tstatus,&tstatus->rhw,tsc,sstatus->size,0);
 		rdamage = (int64)rdamage * ratio / 100 + wd->damage * (10 + tsc->getSCE(SC_CRESCENTELBOW)->val1 * 20 / 10) / 10;
